+++ apps/probability/models/statistic/pooled_two_means_statistic.h
@@ -7,10 +7,10 @@ namespace Probability {
 
 class PooledTwoMeansStatistic : public TwoMeansTStatistic {
 public:
-  float computeDegreesOfFreedom(float s1, int n1, float s2, int n2) override;
-  float computeStandardError(float s1, int n1, float s2, int n2) override;
+  float computeDegreesOfFreedom(float s1, float n1, float s2, float n2) override;
+  float computeStandardError(float s1, float n1, float s2, float n2) override;
   /* Pooled estimate of sigma */
-  float _sp(float s1, int n1, float s2, int n2);
+  float _sp(float s1, float n1, float s2, float n2);
 };
 
 }  // namespace Probability
