+++ apps/probability/models/statistic/two_means_t_statistic.h
@@ -37,8 +37,8 @@ private:
   float _xEstimate(float meanSample1, float meanSample2);
   float computeT(float deltaMean, float meanSample1, float n1, float sigma1, float meanSample2, float n2,
            float sigma2);
-  virtual float computeDegreesOfFreedom(float s1, int n1, float s2, int n2);
-  virtual float computeStandardError(float s1, int n1, float s2, int n2);
+  virtual float computeDegreesOfFreedom(float s1, float n1, float s2, float n2);
+  virtual float computeStandardError(float s1, float n1, float s2, float n2);
 
   float m_params[k_numberOfParams];
 };
